# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"GPU Cluster Platform","version":"v1"},"servers":[{"url":"http://localhost:18080","description":"Generated server url"}],"security":[{"bearer":[]}],"paths":{"/v1/workloads":{"get":{"tags":["platform-controller"],"operationId":"workloads","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}},"post":{"tags":["platform-controller"],"operationId":"create","parameters":[{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Workload"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/workloads/{id}/{action}":{"post":{"tags":["platform-controller"],"operationId":"change","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"action","in":"path","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Change"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/workloads/{id}/revisions":{"post":{"tags":["platform-controller"],"operationId":"revision","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Revision"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/nodes/{name}/{action}":{"post":{"tags":["node-controller"],"operationId":"change_1","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}},{"name":"action","in":"path","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeChange"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/image-uploads":{"post":{"tags":["platform-controller"],"operationId":"upload","parameters":[{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Upload"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/image-uploads/{id}/credentials":{"post":{"tags":["platform-controller"],"operationId":"credentials","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/image-uploads/{id}/complete":{"post":{"tags":["platform-controller"],"operationId":"complete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Complete"}}},"required":true},"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/workloads/{id}":{"get":{"tags":["platform-controller"],"operationId":"workload","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}},"delete":{"tags":["platform-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"If-Match","in":"header","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/operations/{id}":{"get":{"tags":["platform-controller"],"operationId":"operation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/nodes":{"get":{"tags":["node-controller"],"operationId":"list","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/nodes/{name}":{"get":{"tags":["node-controller"],"operationId":"detail","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/images":{"get":{"tags":["platform-controller"],"operationId":"images","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/images/{id}":{"get":{"tags":["platform-controller"],"operationId":"image","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}},"delete":{"tags":["platform-controller"],"operationId":"deleteImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"Idempotency-Key","in":"header","description":"Caller/method/path scoped; retained 7 days","required":true,"schema":{"type":"string","maxLength":128}}],"responses":{"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"},"202":{"description":"Persisted operation accepted; poll Location","content":{"application/json":{"schema":{"type":"object","properties":{"operation_id":{"type":"string"},"target_id":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"description":"Invalid or unknown fields"},"409":{"description":"Idempotency, state or version conflict"},"503":{"description":"Persistence unavailable; operation not accepted"}}}},"/v1/image-uploads/{id}":{"get":{"tags":["platform-controller"],"operationId":"upload_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}},"/v1/capabilities":{"get":{"tags":["platform-controller"],"operationId":"capabilities","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}},"401":{"description":"Missing or invalid authentication"},"403":{"description":"Caller or scope not permitted"}}}}},"components":{"schemas":{"Placement":{"type":"object","properties":{"nodePool":{"type":"string","pattern":"[a-z0-9-]{1,63}"},"allowedNodes":{"type":"array","items":{"type":"string","pattern":"[a-z0-9.-]{1,253}"},"maxItems":64,"minItems":0}}},"Port":{"type":"object","properties":{"name":{"type":"string","minLength":1,"pattern":"[a-z][a-z0-9-]{0,14}"},"port":{"type":"integer","format":"int32","maximum":65535,"minimum":1},"protocol":{"type":"string","minLength":1,"pattern":"TCP|UDP"}},"required":["name","protocol"]},"Probe":{"type":"object","properties":{"type":{"type":"string","minLength":1,"pattern":"http|tcp|process"},"path":{"type":"string","maxLength":256,"minLength":0},"port":{"type":"integer","format":"int32","maximum":65535,"minimum":0}},"required":["type"]},"Resources":{"type":"object","properties":{"gpuPerReplica":{"type":"integer","format":"int32","maximum":8,"minimum":0},"cpuCores":{"type":"number","format":"double","maximum":256,"minimum":0.1},"memoryGib":{"type":"number","format":"double","maximum":2048,"minimum":0.0625}}},"Workload":{"type":"object","properties":{"name":{"type":"string","minLength":1,"pattern":"[a-z][a-z0-9-]{0,49}"},"imageId":{"type":"string","minLength":1},"clusterId":{"type":"string","minLength":1},"type":{"type":"string","minLength":1,"pattern":"service"},"replicas":{"type":"integer","format":"int32","maximum":64,"minimum":1},"resources":{"$ref":"#/components/schemas/Resources"},"placement":{"$ref":"#/components/schemas/Placement"},"command":{"type":"array","items":{"type":"string","maxLength":1024,"minLength":0},"maxItems":32,"minItems":0},"args":{"type":"array","items":{"type":"string","maxLength":4096,"minLength":0},"maxItems":64,"minItems":0},"env":{"type":"object","additionalProperties":{"type":"string"}},"ports":{"type":"array","items":{"$ref":"#/components/schemas/Port"},"maxItems":16,"minItems":0},"readiness":{"$ref":"#/components/schemas/Probe"},"startup":{"$ref":"#/components/schemas/Probe"},"terminationGraceSeconds":{"type":"integer","format":"int32","maximum":600,"minimum":1}},"required":["clusterId","imageId","name","ports","readiness","resources","type"]},"Change":{"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int64","minimum":1},"replicas":{"type":"integer","format":"int32","maximum":64,"minimum":0}}},"Revision":{"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int64","minimum":1},"configuration":{"$ref":"#/components/schemas/Workload"}},"required":["configuration"]},"NodeChange":{"type":"object","properties":{"expectedUid":{"type":"string","maxLength":128,"minLength":0},"expectedResourceVersion":{"type":"string","maxLength":128,"minLength":0}},"required":["expectedResourceVersion","expectedUid"]},"Upload":{"type":"object","properties":{"repository":{"type":"string","minLength":1,"pattern":"[a-z0-9][a-z0-9-]{0,39}"},"tag":{"type":"string","minLength":1,"pattern":"[a-zA-Z0-9_][a-zA-Z0-9_.-]{0,127}"}},"required":["repository","tag"]},"Complete":{"type":"object","properties":{"digest":{"type":"string","pattern":"sha256:[a-f0-9]{64}"}}}},"securitySchemes":{"bearer":{"type":"http","scheme":"bearer"}}}}